% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixelArray_Constructor.R
\name{analyseNwriteOneFixel.lm}
\alias{analyseNwriteOneFixel.lm}
\title{Analyse (fit statistical model) and write the outputs for 1 fixel}
\usage{
analyseNwriteOneFixel.lm(
  i_fixel,
  formula,
  fa,
  phenotypes,
  scalar,
  fn.output.h5,
  analysis_name = "lm",
  var.terms,
  var.model,
  flag_initiate = FALSE,
  overwrite = TRUE,
  results.grp = NULL,
  results.analysis.grp = NULL,
  results_matrix_ds = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{i_fixel}{The i_th fixel, starting from 1, integer. For initiating (flag_initiate = TRUE), use i_fixel=1}

\item{formula}{Formula (passed to `lm()`)}

\item{fa}{FixelArray class}

\item{phenotypes}{The cohort matrix with covariates to be added to the model}

\item{scalar}{The name of the scalar to be analysed fixel-wise}

\item{fn.output.h5}{Opened h5 file (via H5File$new(filename, mode="a"))}

\item{analysis_name}{The subgroup name in results, holding the analysis results}

\item{var.terms}{The list of variables to save for terms (got from lm %>% tidy())}

\item{var.model}{The list of variables to save for the model (got from lm %>% glance())}

\item{flag_initiate}{Whether this is to initiate the new group (TRUE or FALSE) - if this is the first i_fixel, then TRUE.}

\item{overwrite}{Whether to overwrite or not    # TODO: to make this description more clear}

\item{results.grp}{# TODO: to explain}

\item{results.analysis.grp}{}

\item{verbose}{Print progress messages}
}
\description{
Analyse (fit statistical model) and write the outputs for 1 fixel
}
