% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixelArray_Constructor.R
\name{analyseOneFixel.gam}
\alias{analyseOneFixel.gam}
\title{Analyse (fit mgcv::gam()) and write the outputs for 1 fixel}
\usage{
analyseOneFixel.gam(
  i_fixel,
  formula,
  fa,
  phenotypes,
  scalar,
  var.smoothTerms,
  var.parametricTerms,
  var.model,
  flag_initiate = FALSE,
  ...
)
}
\arguments{
\item{i_fixel}{The i_th fixel, starting from 1, integer. For initiating (flag_initiate = TRUE), use i_fixel=1}

\item{formula}{Formula (passed to `mgcv::gam()`)}

\item{fa}{FixelArray class}

\item{phenotypes}{The cohort matrix with covariates to be added to the model}

\item{scalar}{The name of the scalar to be analysed fixel-wise}

\item{var.smoothTerms}{The list of variables to save for smooth terms (got from gam %>% tidy(parametric = FALSE)). Example smooth term: age in formula "outcome ~ s(age)".}

\item{var.parametricTerms}{The list of variables to save for parametric terms (got from gam %>% tidy(parametric = TRUE)). Example parametric term: sex in formula "outcome ~ s(age) + sex"}

\item{var.model}{The list of variables to save for the model (got from gam %>% glance() and gam %>% summary())}

\item{flag_initiate}{Whether this is to initiate the new group (TRUE or FALSE) - if this is the first i_fixel, then TRUE and it will return column names.}
}
\value{
if flag_initiate==TRUE, returns column names & list of terms of final results; if flag_initiate==FALSE, returns the final results for a fixel
}
\description{
Analyse (fit mgcv::gam()) and write the outputs for 1 fixel
}
