% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixelArray_Constructor.R
\name{analyseOneFixel.lm}
\alias{analyseOneFixel.lm}
\title{Analyse (fit statistical model) and write the outputs for 1 fixel}
\usage{
analyseOneFixel.lm(
  i_fixel,
  formula,
  fa,
  phenotypes,
  scalar,
  var.terms,
  var.model,
  flag_initiate = FALSE,
  ...
)
}
\arguments{
\item{i_fixel}{The i_th fixel, starting from 1, integer. For initiating (flag_initiate = TRUE), use i_fixel=1}

\item{formula}{Formula (passed to `lm()`)}

\item{fa}{FixelArray class}

\item{phenotypes}{The cohort matrix with covariates to be added to the model}

\item{scalar}{The name of the scalar to be analysed fixel-wise}

\item{var.terms}{The list of variables to save for terms (got from lm %>% tidy())}

\item{var.model}{The list of variables to save for the model (got from lm %>% glance())}

\item{flag_initiate}{Whether this is to initiate the new group (TRUE or FALSE) - if this is the first i_fixel, then TRUE and it will return column names.}
}
\value{
if flag_initiate==TRUE, returns column names of final results; if flag_initiate==FALSE, returns the final results for a fixel
}
\description{
Analyse (fit statistical model) and write the outputs for 1 fixel
}
