% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_Constructor.R
\name{analyseOneGrid.lm}
\alias{analyseOneGrid.lm}
\title{Fit linear model for one grid.}
\usage{
analyseOneGrid.lm(
  i_grid,
  formula,
  modelarray,
  phenotypes,
  scalar,
  var.terms,
  var.model,
  flag_initiate = FALSE,
  ...
)
}
\arguments{
\item{i_grid}{An integer, the i_th grid, starting from 1. For initiating (flag_initiate = TRUE), use i_grid=1}

\item{formula}{Formula (passed to `stats::lm()`)}

\item{modelarray}{ModelArray class}

\item{phenotypes}{A data.frame of the cohort with columns of independent variables and covariates to be added to the model.}

\item{scalar}{A character. The name of the grid-wise scalar to be analysed}

\item{var.terms}{A list of characters. The list of variables to save for terms (got from `broom::tidy()`).}

\item{var.model}{A list of characters. The list of variables to save for the model (got from `broom::glance()`).}

\item{flag_initiate}{TRUE or FALSE, Whether this is to initiate the new analysis. If TRUE, it will return column names etc to be used for initiating data.frame; if FALSE, it will return the list of requested statistic values.}

\item{...}{Additional arguments for `stats::lm()`}
}
\value{
If flag_initiate==TRUE, returns column names, and list of term names of final results; if flag_initiate==FALSE, it will return the list of requested statistic values for a grid.
}
\description{
Fit linear model for one grid.
}
