% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_S4Methods.R
\name{analysisNames}
\alias{analysisNames}
\alias{analysisNames,ModelArray-method}
\title{Names of analyses in a ModelArray}
\usage{
analysisNames(x)

\S4method{analysisNames}{ModelArray}(x)
}
\arguments{
\item{x}{A ModelArray object}
}
\value{
Character vector of analysis names
}
\description{
Names of analyses in a ModelArray
}
