% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{check_validity_correctPValue}
\alias{check_validity_correctPValue}
\title{Check if the list of p-value correction methods are valid for a specific type of term/model. 
Can be used for any statistical model. As long as the p.value to be correct is named as "p.value".}
\usage{
check_validity_correctPValue(
  correct.list,
  name.correct.list,
  var.list,
  name.var.list
)
}
\arguments{
\item{correct.list}{The list of correction methods for this type of term/model}

\item{name.correct.list}{The name of the list of correction methods for this type of term/model}

\item{var.list}{The list of statistics to be saved for this type of term/model}

\item{name.var.list}{The name of the list of statistics to be saved for this type of term/model}
}
\description{
Check if the list of p-value correction methods are valid for a specific type of term/model. 
Can be used for any statistical model. As long as the p.value to be correct is named as "p.value".
}
