% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{checker_gam_s}
\alias{checker_gam_s}
\title{Print out important arguments in smooth terms s() in mgcv::gam() formula
ref: https://www.rdocumentation.org/packages/mgcv/versions/1.8-38/topics/s
TODO: finish the description}
\usage{
checker_gam_s(ofInterest)
}
\arguments{
\item{ofInterest}{got via: gam.formula.breakdown <- mgcv::interpret.gam(formula); ofInterest <- gam.formula.breakdown$smooth.spec[[i]]}
}
\description{
Print out important arguments in smooth terms s() in mgcv::gam() formula
ref: https://www.rdocumentation.org/packages/mgcv/versions/1.8-38/topics/s
TODO: finish the description
}
