% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{checker_gam_t}
\alias{checker_gam_t}
\title{Print out important arguments in smooth term te() or ti() or t2() in mgcv::gam() formula
Why a separate function is needed for t(), cannot using s(): in ofInterest, "fx" is "fx" for t(), but "fixed" for s() - so they are different.
ref: https://www.rdocumentation.org/packages/mgcv/versions/1.8-38/topics/te or /t2()
TODO: finish the description}
\usage{
checker_gam_t(FUN, ofInterest)
}
\arguments{
\item{FUN}{could be mgcv::te(), ti() or t2()}

\item{ofInterest}{got via: gam.formula.breakdown <- mgcv::interpret.gam(formula); ofInterest <- gam.formula.breakdown$smooth.spec[[i]]}
}
\description{
Print out important arguments in smooth term te() or ti() or t2() in mgcv::gam() formula
Why a separate function is needed for t(), cannot using s(): in ofInterest, "fx" is "fx" for t(), but "fixed" for s() - so they are different.
ref: https://www.rdocumentation.org/packages/mgcv/versions/1.8-38/topics/te or /t2()
TODO: finish the description
}
