% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_S4Methods.R
\name{elementMetadata}
\alias{elementMetadata}
\alias{elementMetadata,ModelArray-method}
\title{Element metadata from a ModelArray}
\usage{
elementMetadata(x)

\S4method{elementMetadata}{ModelArray}(x)
}
\arguments{
\item{x}{A ModelArray object}
}
\value{
A matrix or data.frame of element metadata, or NULL
}
\description{
Reads element metadata (e.g., greyordinates for cifti data) from the h5 file
if present. Returns NULL if no element metadata is found.
}
