% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{FixelArray.lm}
\alias{FixelArray.lm}
\title{Run a linear model at each fixel location, write out each result just after the model fitting}
\usage{
FixelArray.lm(
  formula,
  data,
  phenotypes,
  scalar,
  fixel.subset = NULL,
  full.outputs = FALSE,
  var.terms = c("estimate", "statistic", "p.value"),
  var.model = c("adj.r.squared", "p.value"),
  correct.p.value.terms = "none",
  correct.p.value.model = "none",
  verbose = TRUE,
  pbar = TRUE,
  n_cores = 1,
  ...
)
}
\arguments{
\item{formula}{Formula (passed to `lm()`)}

\item{data}{FixelArray class}

\item{phenotypes}{The cohort matrix with covariates to be added to the model}

\item{scalar}{The name of the scalar to be analysed fixel-wise}

\item{fixel.subset}{The subset of fixel ids you want to run. Integers. First id starts from 1.}

\item{full.outputs}{Whether to return full set of outputs (TRUE or FALSE). If FALSE, it will only return those listed in var.terms and var.model; if TRUE, arguments var.terms and var.model will be ignored.}

\item{var.terms}{The list of variables to save for terms (got from lm %>% tidy())}

\item{var.model}{The list of variables to save for the model (got from lm %>% glance())}

\item{correct.p.value.terms}{To perform and add a column for p.value correction for each term. Currently "fdr", "bonferroni" are supported, can be more than one. Turn it off by setting to "none".}

\item{correct.p.value.model}{To perform and add a column for p.value correction for the model. Currently "fdr", "bonferroni" are supported, can be more than one. Turn it off by setting to "none".}

\item{verbose}{Print verbose message or not}

\item{pbar}{Print progress bar}

\item{n_cores}{The number of cores to run on}
}
\description{
Run a linear model at each fixel location, write out each result just after the model fitting
}
