% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixelArray_Constructor.R
\name{FixelArray}
\alias{FixelArray}
\title{An S4 class to represent a bank account.}
\usage{
FixelArray(filepath, scalar_types = c("FD"), analysis_names = c("myAnalysis"))

FixelArray(filepath, scalar_types = c("FD"), analysis_names = c("myAnalysis"))
}
\arguments{
\item{filepath}{file}

\item{scalar_types}{expected scalars}

\item{analysis_names}{the subfolder names for results in .h5 file}
}
\value{
FixelArray object
}
\description{
An S4 class to represent a bank account.

Load fixel data output from mrtrix as an h5 file into R as a FixelArray object
Tips for debugging: 
if you run into this error: "Error in h(simpleError(msg, call)) : error in evaluating the argument 'seed' in selecting a method for function 'DelayedArray': HDF5. Symbol table. Can't open object." Then please check if you give correct "scalar_types" - check via h5ls(filename_for_h5)
TODO: IN THE FUTURE, THE SCALAR_TYPES AND ANALYSIS_NAMES SHOULD BE AUTOMATICALLY DETECTED!
}
\section{Slots}{

\describe{
\item{\code{fixels}}{A DelayedArray object of fixel data}

\item{\code{voxels}}{A DelayedArray object of voxel indeces}

\item{\code{results}}{An h5 group of FixelArray analysis outputs}

\item{\code{subjects}}{A list of subject labels}

\item{\code{scalars}}{A list of scalars measured by the fixels}

\item{\code{path}}{Path to the h5 file on disk}
}}

