% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{generator_gamFormula_factorXsmooth}
\alias{generator_gamFormula_factorXsmooth}
\title{Generate GAM formula with interaction term: factor-smooth interaction
Example 
`factor.var` and `smooth.var` should come from data.frame `phenotypes`
TODO: finish description}
\usage{
generator_gamFormula_factorXsmooth(
  response.var,
  factor.var,
  smooth.var,
  phenotypes,
  reference.group = NULL,
  prefix.ordered.factor = "o",
  fx = TRUE,
  k = -1
)
}
\arguments{
\item{response.var}{character class, the variable name for response}

\item{factor.var}{character class, the variable name for factor. It should be an ordered factor. If not, it will generate it as a new column in `phenotypes`, which requires `reference.group`.}

\item{smooth.var}{character class, the variable name in smooth term as main effect}

\item{phenotypes}{data.frame class, the cohort matrix with covariates to be added to the model}

\item{reference.group}{character class, the reference group for ordered factor of `factor.var`; required when `factor.var` in `phenotypes` is not an ordered factor.}

\item{prefix.ordered.factor}{character class, the prefix for ordered factor; required when `factor.var` in `phenotypes` is not an ordered factor.}

\item{fx}{TRUE or FALSE, to be used in smooth term s(). Recommend TRUE.}

\item{k}{integer, to be used in smooth term s(). Default is -1 as in mgcv::s()}
}
\value{
a list, including: 1) formula generated; 2) data.frame phenotypes - updated if argument factor.var is not an ordered factor
}
\description{
Generate GAM formula with interaction term: factor-smooth interaction
Example 
`factor.var` and `smooth.var` should come from data.frame `phenotypes`
TODO: finish description
}
