% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_Constructor.R
\name{ModelArray-class}
\alias{ModelArray-class}
\title{ModelArray class}
\description{
An S4 class to represent element-wise scalar data and statistics.
}
\section{Slots}{

\describe{
\item{\code{sources}}{A list of source filenames}

\item{\code{scalars}}{A list of element-wise scalar matrices}

\item{\code{results}}{A list of statistical result matrices}

\item{\code{path}}{Path to the h5 file on disk}
}}

\keyword{internal}
