% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_Constructor.R
\name{ModelArray}
\alias{ModelArray}
\title{Construct a ModelArray object}
\usage{
ModelArray(filepath, scalar_types = c("FD"), analysis_names = character(0))
}
\arguments{
\item{filepath}{Path to an .h5 file}

\item{scalar_types}{Expected scalars}

\item{analysis_names}{The subfolder names for results in the .h5 file.
If empty (default), results are not read.}
}
\value{
A `ModelArray` object
}
\description{
Load element-wise data from an .h5 file as a `ModelArray` object.
}
