% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{ModelArray.t.test}
\alias{ModelArray.t.test}
\title{Run a t.test at each element location}
\usage{
ModelArray.t.test(
  formula,
  data,
  phenotypes,
  scalar,
  verbose = TRUE,
  idx = NULL,
  pbar = TRUE,
  n_cores = 1,
  ...
)
}
\arguments{
\item{formula}{Formula (passed to `lm()`)}

\item{data}{ModelArray dataset}

\item{phenotypes}{The cohort file with covariates to be added to the model}

\item{scalar}{The name of the scalar to be analysed element-wise}

\item{verbose}{Print progress messages}

\item{pbar}{Print progress bar}

\item{n_cores}{The number of cores to run on}

\item{subset}{A vector of element IDs to subset}
}
\value{
Tibble with the summarised model statistics at each element location
}
\description{
Run a t.test at each element location
}
