% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_S4Methods.R
\name{nElements}
\alias{nElements}
\alias{nElements,ModelArray-method}
\title{Number of elements in a ModelArray}
\usage{
nElements(x, scalar = NULL)

\S4method{nElements}{ModelArray}(x, scalar = NULL)
}
\arguments{
\item{x}{A ModelArray object}

\item{scalar}{Optional scalar name. Defaults to the first scalar.}
}
\value{
Integer, number of elements (rows)
}
\description{
Number of elements in a ModelArray
}
