% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_Constructor.R
\name{numElementsTotal}
\alias{numElementsTotal}
\title{Number of elements in ModelArray}
\usage{
numElementsTotal(modelarray, scalar_name = "FD")
}
\arguments{
\item{modelarray}{ModelArray class}

\item{scalar_name}{A character, the scalar name (one of the existing scalar in \code{modelarray})}
}
\value{
numElementsTotal number of elements in ModelArray, for this specific scalar
}
\description{
Returns the number of elements in ModelArray, for a specific scalar
}
