% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_Constructor.R
\name{numGridsTotal}
\alias{numGridsTotal}
\title{Number of grids in ModelArray}
\usage{
numGridsTotal(modelarray, scalar_name = "FD")
}
\arguments{
\item{modelarray}{ModelArray class}

\item{scalar_name}{A character, the scalar name (one of the existing scalar in \code{modelarray})}
}
\value{
numGridsTotal number of grids in ModelArray, for this specific scalar
}
\description{
Returns the number of grids in ModelArray, for a specific scalar
}
