% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_S4Methods.R
\name{results,ModelArray-method}
\alias{results,ModelArray-method}
\title{Statistical results of an ModelArray object}
\usage{
\S4method{results}{ModelArray}(x, ...)
}
\arguments{
\item{x}{An ModelArray object}

\item{...}{Additional arguments. Currently accept analysis name (a character)}
}
\value{
Statistical results in this ModelArray object
}
\description{
Statistical results of an ModelArray object
}
