% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_S4Methods.R
\name{results}
\alias{results}
\alias{results,ModelArray-method}
\title{Statistical results of a ModelArray object}
\usage{
results(x, ...)

\S4method{results}{ModelArray}(x, ...)
}
\arguments{
\item{x}{A ModelArray object}

\item{...}{Additional arguments. Currently accepts an analysis name (character).}
}
\value{
Statistical results in this ModelArray object
}
\description{
Statistical results of a ModelArray object
}
