% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_S4Methods.R
\name{scalarNames}
\alias{scalarNames}
\alias{scalarNames,ModelArray-method}
\title{Names of scalars in a ModelArray}
\usage{
scalarNames(x)

\S4method{scalarNames}{ModelArray}(x)
}
\arguments{
\item{x}{A ModelArray object}
}
\value{
Character vector of scalar names
}
\description{
Names of scalars in a ModelArray
}
