% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_S4Methods.R
\name{scalars}
\alias{scalars}
\alias{scalars,ModelArray-method}
\title{Element-wise scalar data of a ModelArray object}
\usage{
scalars(x, ...)

\S4method{scalars}{ModelArray}(x, ...)
}
\arguments{
\item{x}{A ModelArray object}

\item{...}{Additional arguments. Currently accepts a scalar name (character).}
}
\value{
A matrix of element-wise scalar data: elements (row) by source files (column).
}
\description{
Element-wise scalar data of a ModelArray object
}
