% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_S4Methods.R
\name{sources,ModelArray-method}
\alias{sources,ModelArray-method}
\title{Source filenames of an ModelArray object}
\usage{
\S4method{sources}{ModelArray}(x)
}
\arguments{
\item{x}{An ModelArray object}
}
\value{
A list of source filenames
}
\description{
Source filenames of an ModelArray object
}
