% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_S4Methods.R
\name{sources}
\alias{sources}
\alias{sources,ModelArray-method}
\title{Source filenames of a ModelArray object}
\usage{
sources(x)

\S4method{sources}{ModelArray}(x)
}
\arguments{
\item{x}{A ModelArray object}
}
\value{
A list of source filenames
}
\description{
Source filenames of a ModelArray object
}
