% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixelArray_Constructor.R
\name{writeResults.old}
\alias{writeResults.old}
\title{Write outputs from fixel-based analysis out to the h5 file. Write one results (i.e. for one analysis) at a time. This is ".old": for writing results with multiple rows for one fixel}
\usage{
writeResults.old(fa, data, analysis_name = "myAnalysis", flag_overwrite = TRUE)
}
\arguments{
\item{fa}{FixelArray object}

\item{data}{A data.frame object with model results at each fixel}

\item{analysis_name}{The subfolder name in results, holding the analysis results}

\item{flag_overwrite}{If same analysis_name exists, whether overwrite or not}
}
\description{
Write outputs from fixel-based analysis out to the h5 file. Write one results (i.e. for one analysis) at a time. This is ".old": for writing results with multiple rows for one fixel
}
