% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_Constructor.R
\name{analyseOneElement.gam}
\alias{analyseOneElement.gam}
\title{Fit GAM for one element}
\usage{
analyseOneElement.gam(
  i_element,
  formula,
  modelarray,
  phenotypes,
  scalar,
  var.smoothTerms,
  var.parametricTerms,
  var.model,
  flag_initiate = FALSE,
  flag_sse = FALSE,
  ...
)
}
\arguments{
\item{i_element}{An integer, the i_th element, starting from 1. For initiating (flag_initiate = TRUE), use i_element=1}

\item{formula}{A formula (passed to `mgcv::gam()`)}

\item{modelarray}{ModelArray class}

\item{phenotypes}{A data.frame of the cohort with columns of independent variables and covariates to be added to the model}

\item{scalar}{A character. The name of the element-wise scalar to be analysed}

\item{var.smoothTerms}{The list of variables to save for smooth terms (got from broom::tidy(parametric = FALSE)). Example smooth term: age in formula "outcome ~ s(age)".}

\item{var.parametricTerms}{The list of variables to save for parametric terms (got from broom::tidy(parametric = TRUE)). Example parametric term: sex in formula "outcome ~ s(age) + sex".}

\item{var.model}{The list of variables to save for the model (got from broom::glance() and summary()).}

\item{flag_initiate}{TRUE or FALSE, Whether this is to initiate the new analysis. If TRUE, it will return column names etc to be used for initiating data.frame; if FALSE, it will return the list of requested statistic values.}

\item{flag_sse}{TRUE or FALSE, Whether to calculate SSE (sum of squared error) for the model (`model.sse`). SSE is needed for calculating partial R-squared.}

\item{...}{Additional arguments for `mgcv::gam()`}
}
\value{
If flag_initiate==TRUE, returns column names, list of term names of final results, and attr.name of sp.criterion; if flag_initiate==FALSE, it will return the list of requested statistic values for a element.
}
\description{
`analyseOneElement.gam` fits a GAM model for one element data, and returns requested model statistics.
}
\details{
`ModelArray.gam` iteratively calls this function to get statistics for all requested elements.
}
