% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_Constructor.R
\name{analyseOneElement.lm}
\alias{analyseOneElement.lm}
\title{Fit linear model for one element.}
\usage{
analyseOneElement.lm(
  i_element,
  formula,
  modelarray,
  phenotypes,
  scalar,
  var.terms,
  var.model,
  flag_initiate = FALSE,
  ...
)
}
\arguments{
\item{i_element}{An integer, the i_th element, starting from 1. For initiating (flag_initiate = TRUE), use i_element=1}

\item{formula}{Formula (passed to `stats::lm()`)}

\item{modelarray}{ModelArray class}

\item{phenotypes}{A data.frame of the cohort with columns of independent variables and covariates to be added to the model.}

\item{scalar}{A character. The name of the element-wise scalar to be analysed}

\item{var.terms}{A list of characters. The list of variables to save for terms (got from `broom::tidy()`).}

\item{var.model}{A list of characters. The list of variables to save for the model (got from `broom::glance()`).}

\item{flag_initiate}{TRUE or FALSE, Whether this is to initiate the new analysis. If TRUE, it will return column names etc to be used for initiating data.frame; if FALSE, it will return the list of requested statistic values.}

\item{...}{Additional arguments for `stats::lm()`}
}
\value{
If flag_initiate==TRUE, returns column names, and list of term names of final results; if flag_initiate==FALSE, it will return the list of requested statistic values for a element.
}
\description{
`analyseOneElement.lm` fits a linear model for one element data, and returns requested model statistics.
}
\details{
`ModelArray.lm` iteratively calls this function to get statistics for all requested elements.
}
