% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_Constructor.R
\name{analyseOneFixel.gam}
\alias{analyseOneFixel.gam}
\title{Fit GAM for one fixel data}
\usage{
analyseOneFixel.gam(
  i_fixel,
  formula,
  modelarray,
  phenotypes,
  scalar,
  var.smoothTerms,
  var.parametricTerms,
  var.model,
  flag_initiate = FALSE,
  ...
)
}
\arguments{
\item{i_fixel}{An integer, the i_th fixel, starting from 1. For initiating (flag_initiate = TRUE), use i_fixel=1}

\item{formula}{A formula (passed to `mgcv::gam()`)}

\item{modelarray}{ModelArray class}

\item{phenotypes}{A data.frame of the cohort with columns of independent variables and covariates to be added to the model}

\item{scalar}{The name of the scalar to be analysed fixel-wise}

\item{var.smoothTerms}{The list of variables to save for smooth terms (got from broom::tidy(parametric = FALSE)). Example smooth term: age in formula "outcome ~ s(age)".}

\item{var.parametricTerms}{The list of variables to save for parametric terms (got from broom::tidy(parametric = TRUE)). Example parametric term: sex in formula "outcome ~ s(age) + sex".}

\item{var.model}{The list of variables to save for the model (got from broom::glance() and summary()).}

\item{flag_initiate}{TRUE or FALSE, Whether this is to initiate the new analysis. If TRUE, it will return column names to be used for initiating data.frame; if FALSE, it will return the list of requested statistic values.}

\item{...}{Arguments for `mgcv::gam()`}
}
\value{
If flag_initiate==TRUE, returns column names ,and list of term names of final results and attr.name of sp.criterion; if flag_initiate==FALSE, it will return the list of requested statistic values for a fixel.
}
\description{
`analyseOneFixel.gam` fits a GAM model for one fixel data, and returns requested model statistics.
}
\details{
`ModelArray.gam` iteratively calls this function to get statistics for all requested fixels.
}
