% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{FixelArray.lm}
\alias{FixelArray.lm}
\title{Run a linear model at each fixel location, write out each result just after the model fitting}
\usage{
FixelArray.lm(
  formula,
  data,
  phenotypes,
  scalar,
  fixel.subset = NULL,
  var.terms = c("estimate", "p.value"),
  var.model = c("r.squared", "p.value"),
  overwrite = TRUE,
  verbose = TRUE,
  pbar = TRUE,
  n_cores = 1,
  ...
)
}
\description{
Run a linear model at each fixel location, write out each result just after the model fitting
}
