% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixelArray.R
\name{FixelArray}
\alias{FixelArray}
\title{Load fixel data output from mrtrix as an h5 file into R as a FixelArray object}
\usage{
FixelArray(filepath, scalar_types = c("FD"))
}
\arguments{
\item{filepath}{file}

\item{scalar_types}{expected scalars}
}
\value{
FixelArray object
}
\description{
Load fixel data output from mrtrix as an h5 file into R as a FixelArray object
}
