% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_gamFormula_contIx}
\alias{gen_gamFormula_contIx}
\title{Generate GAM formula with continuous*continuous interaction}
\usage{
gen_gamFormula_contIx(response.var, cont1.var, cont2.var, fx = TRUE, k = NULL)
}
\arguments{
\item{response.var}{character class, the variable name for response}

\item{cont1.var}{character class, the name of the first continuous variable}

\item{cont2.var}{character class, the name of the second continuous variable}

\item{fx}{TRUE or FALSE, to be used in smooth term s(). Recommend TRUE.}

\item{k}{integer, to be used in smooth term including the interaction term.
If NULL (no entry), will use default value as in mgcv::s()}
}
\value{
The formula generated
}
\description{
This function will generate a formula in the following format: \code{y ~ ti(x) + ti(z) + ti(x,z)},
where \code{y} is \code{response.var}, \code{x} is \code{cont1.var}, and \code{z} is \code{cont2.var}.
The formula generated could be further modified, e.g. adding covariates.
}
