% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_gamFormula_fxSmooth}
\alias{gen_gamFormula_fxSmooth}
\title{Generate GAM formula with factor-smooth interaction}
\usage{
gen_gamFormula_fxSmooth(
  response.var,
  factor.var,
  smooth.var,
  phenotypes,
  reference.group = NULL,
  prefix.ordered.factor = "o",
  fx = TRUE,
  k = NULL
)
}
\arguments{
\item{response.var}{character class, the variable name for response}

\item{factor.var}{character class, the variable name for factor. It should be an ordered factor.
If not, it will generate it as a new column in `phenotypes`, which requires `reference.group`.}

\item{smooth.var}{character class, the variable name in smooth term as main effect}

\item{phenotypes}{data.frame class, the cohort matrix with columns of independent variables (including
\code{factor.var} and \code{smooth.var}) to be added to the model}

\item{reference.group}{character class, the reference group for ordered factor of `factor.var`; required when
`factor.var` in `phenotypes` is not an ordered factor.}

\item{prefix.ordered.factor}{character class, the prefix for ordered factor; required when `factor.var` in
`phenotypes` is not an ordered factor.}

\item{fx}{TRUE or FALSE, to be used in smooth term s(). Recommend TRUE.}

\item{k}{integer, to be used in smooth term including the interaction term. If NULL (no entry), will use default
value as in mgcv::s()}
}
\value{
a list, including: 1) formula generated; 2) data.frame phenotypes - updated if argument factor.var
is not an ordered factor
}
\description{
This function will generate a formula in the following format:
\code{y ~ orderedFactor + s(x) + s(x, by=orderedFactor)},
where \code{y} is \code{response.var}, \code{x} is \code{smooth.var},
and \code{orderedFactor} is \code{factor.var} - see \code{factor.var} for more.
The formula generated could be further modified, e.g. adding covariates.
}
