% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{ModelArray.lm}
\alias{ModelArray.lm}
\title{Fit linear model for grid-wise data}
\usage{
ModelArray.lm(
  formula,
  data,
  phenotypes,
  scalar,
  grid.subset = NULL,
  full.outputs = FALSE,
  colname.subjid = "subject_id",
  var.terms = c("estimate", "statistic", "p.value"),
  var.model = c("adj.r.squared", "p.value"),
  correct.p.value.terms = "none",
  correct.p.value.model = "none",
  verbose = TRUE,
  pbar = TRUE,
  n_cores = 1,
  ...
)
}
\arguments{
\item{formula}{Formula (passed to `stats::lm()`)}

\item{data}{ModelArray class}

\item{phenotypes}{A data.frame of the cohort with columns of independent variables and covariates to be added to the model. It should contains a column of subject IDs that matches to that in \code{data}.}

\item{scalar}{A character. The name of the grid-wise scalar to be analysed}

\item{grid.subset}{A list of positive integers (min = 1, max = number of grids). The subset of grids you want to run.}

\item{full.outputs}{TRUE or FALSE, Whether to return full set of outputs. If FALSE, it will only return those listed in arguments \code{var.*}; if TRUE, arguments \code{var.*} will be ignored.}

\item{colname.subjid}{A character, the column name in \code{phenotypes} of subject IDs. This column will be used for sanity check for matching of subject list in \code{data}.}

\item{var.terms}{A list of characters. The list of variables to save for terms (got from `broom::tidy()`). See "Details" section for more.}

\item{var.model}{A list of characters. The list of variables to save for the model (got from `broom::glance()`). See "Details" section for more.}

\item{correct.p.value.terms}{A list of characters. To perform and add a column for p.value correction for each term. See "Details" section for more.}

\item{correct.p.value.model}{A list of characters. To perform and add a column for p.value correction for the model. See "Details" section for more.}

\item{verbose}{TRUE or FALSE, to print verbose message or not}

\item{pbar}{TRUE or FALSE, to print progress bar or not}

\item{n_cores}{Positive integer, The number of CPU cores to run with}

\item{...}{Additional arguments for `stats::lm()`}
}
\value{
Tibble with the summarized model statistics for all grids requested
}
\description{
`ModelArray.lm` fits linear model (`stats::lm()`) for each of grids requested, and returns a tibble dataframe of requested model statistics.
}
\details{
You may request returning specific statistical variables by setting \code{var.*}, or you can get all by setting \code{full.outputs=TRUE}. 
Note that statistics covered by \code{full.outputs} or \code{var.*} are the ones from broom::tidy() and broom::glance() only, and do not include corrected p values.
List of acceptable statistic names for each of \code{var.*}:
\itemize{
  \item \code{var.terms}: c("estimate","std.error","statistic","p.value"); For interpretation please see `broom:tidy()`.
  \item \code{var.model}: c("r.squared", "adj.r.squared", "sigma", "statistic", "p.value", "df", "logLik", "AIC", "BIC", "deviance", "df.residual", "nobs"); For interpretation please see `broom::glance()`.
}
For p-value corrections (arguments \code{correct.p.value.*}), supported methods include all methods in `p.adjust.methods` except "none". Can be more than one method. Turn it off by setting to "none".
}
