% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_Constructor.R
\name{ModelArray}
\alias{ModelArray}
\title{An S4 class to represent element-wise scalar data and statistics.}
\usage{
ModelArray(filepath, scalar_types = c("FD"), analysis_names = character(0))

ModelArray(filepath, scalar_types = c("FD"), analysis_names = character(0))
}
\arguments{
\item{filepath}{file}

\item{scalar_types}{expected scalars}

\item{analysis_names}{the subfolder names for results in .h5 file. If empty
(default), results are not read.}
}
\value{
ModelArray object
}
\description{
An S4 class to represent element-wise scalar data and statistics.

Load element-wise data from .h5 file as an ModelArray object
}
\details{
:
Tips for debugging:
if you run into this error: "Error in h(simpleError(msg, call)) :
error in evaluating the argument 'seed' in selecting a method for
function 'DelayedArray': HDF5. Symbol table. Can't open object."
Then please check if you give correct "scalar_types" - check via
rhdf5::h5ls(filename_for_h5)
}
\section{Slots}{

\describe{
\item{\code{sources}}{A list of source filenames}

\item{\code{scalars}}{A list of element-wise scalar matrix}

\item{\code{results}}{A list of statistical result matrix}

\item{\code{path}}{Path to the h5 file on disk}
}}

