% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_S4Methods.R
\name{nElements}
\alias{nElements}
\alias{nElements,ModelArray-method}
\title{Number of elements in a ModelArray}
\usage{
nElements(x, scalar = NULL)

\S4method{nElements}{ModelArray}(x, scalar = NULL)
}
\arguments{
\item{x}{A \linkS4class{ModelArray} object.}

\item{scalar}{Optional character string. Name of the scalar to query.
Defaults to the first scalar in \code{names(scalars(x))}.}
}
\value{
Integer. The number of elements (rows) in the scalar matrix.
}
\description{
Returns the number of elements (e.g., fixels or voxels) for a given scalar
in a \linkS4class{ModelArray}. This is the row count of the scalar matrix.
}
\examples{
\dontrun{
ma <- ModelArray("data.h5", scalar_types = c("FD"))
nElements(ma)
nElements(ma, "FD")
}

}
\seealso{
\code{\link{nInputFiles}}, \code{\link{numElementsTotal}},
\code{\link{scalars}}, \code{\link{scalarNames}}
}
