% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_S4Methods.R
\name{nInputFiles}
\alias{nInputFiles}
\alias{nInputFiles,ModelArray-method}
\title{Number of input files in a ModelArray}
\usage{
nInputFiles(x, scalar = NULL)

\S4method{nInputFiles}{ModelArray}(x, scalar = NULL)
}
\arguments{
\item{x}{A ModelArray object}

\item{scalar}{Optional scalar name. Defaults to the first scalar.}
}
\value{
Integer, number of input files (columns)
}
\description{
Number of input files in a ModelArray
}
