% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_Constructor.R,
%   R/ModelArray_S4Methods.R
\name{results,ModelArray-method}
\alias{results,ModelArray-method}
\title{Access the results slot of a ModelArray object}
\usage{
\S4method{results}{ModelArray}(x, ...)

\S4method{results}{ModelArray}(x, ...)
}
\arguments{
\item{x}{An ModelArray object}

\item{...}{Additional arguments. Currently accept analysis name (a character)}
}
\value{
The results slot of the ModelArray object

Statistical results in this ModelArray object
}
\description{
Method for accessing the results slot of a ModelArray object.
}
\keyword{internal}
