% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_Constructor.R
\name{results}
\alias{results}
\title{Access the results slot of an object}
\usage{
results(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments passed to methods}
}
\value{
The results slot of the object
}
\description{
Generic function to access the results slot of an object.
}
\keyword{internal}
