% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_S4Methods.R
\name{scalarNames}
\alias{scalarNames}
\alias{scalarNames,ModelArray-method}
\title{Names of scalars in a ModelArray}
\usage{
scalarNames(x)

\S4method{scalarNames}{ModelArray}(x)
}
\arguments{
\item{x}{A \linkS4class{ModelArray} object.}
}
\value{
Character vector of scalar names.
}
\description{
Returns the names of all scalar datasets loaded in a
\linkS4class{ModelArray} (e.g. \code{"FD"}, \code{"FC"}, \code{"log_FC"}).
}
\examples{
\dontrun{
ma <- ModelArray("data.h5", scalar_types = c("FD", "FC"))
scalarNames(ma)   # c("FD", "FC")
}

}
\seealso{
\code{\link{scalars}}, \code{\link{analysisNames}},
\code{\link{nElements}}
}
