% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_Constructor.R,
%   R/ModelArray_S4Methods.R
\name{scalars,ModelArray-method}
\alias{scalars,ModelArray-method}
\title{Access the scalars slot of a ModelArray object}
\usage{
\S4method{scalars}{ModelArray}(x, ...)

\S4method{scalars}{ModelArray}(x, ...)
}
\arguments{
\item{x}{An ModelArray object}

\item{...}{Additional arguments. Currently accept scalar name (a character)}
}
\value{
The scalars slot of the ModelArray object

A matrix of element-wise scalar data: elements (row) by source files (column).
}
\description{
Method for accessing the scalars slot of a ModelArray object.
}
\keyword{internal}
