% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_Constructor.R
\name{scalars}
\alias{scalars}
\title{Access the scalars slot of an object}
\usage{
scalars(x)
}
\arguments{
\item{x}{An object}
}
\value{
The scalars slot of the object
}
\description{
Generic function to access the scalars slot of an object.
}
