% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_Constructor.R,
%   R/ModelArray_S4Methods.R
\name{sources,ModelArray-method}
\alias{sources,ModelArray-method}
\title{Access the sources slot of a ModelArray object}
\usage{
\S4method{sources}{ModelArray}(x)

\S4method{sources}{ModelArray}(x)
}
\arguments{
\item{x}{An ModelArray object}
}
\value{
The sources slot of the ModelArray object

A list of source filenames
}
\description{
Method for accessing the sources slot of a ModelArray object.
}
