% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_Constructor.R
\name{sources}
\alias{sources}
\title{Access the sources slot of an object}
\usage{
sources(x)
}
\arguments{
\item{x}{An object}
}
\value{
The sources slot of the object
}
\description{
Generic function to access the sources slot of an object.
}
\keyword{internal}
