% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixelArray.R
\name{writeResults}
\alias{writeResults}
\title{Write outputs from fixel-based analysis out to the h5 file}
\usage{
writeResults(fa, data, name = "results")
}
\arguments{
\item{fa}{FixelArray object}

\item{data}{A data.frame object with model results at each fixel}

\item{name}{Name assigned to the results table in the h5 file}
}
\description{
Write outputs from fixel-based analysis out to the h5 file
}
