% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixelArray_Constructor.R
\name{writeResults}
\alias{writeResults}
\title{Write outputs from fixel-based analysis out to the h5 file. Write one results (i.e. for one analysis) at a time. This is ".enh": 1) change to hdf5r; 2) write results with only one row for one fixel
debug tip: For "Error in H5File.open(filename, mode, file_create_pl, file_access_pl)", check if there is message 'No such file or directory'. Try absolute .h5 filename.}
\usage{
writeResults(
  fn.output,
  df.output,
  analysis_name = "myAnalysis",
  overwrite = TRUE
)
}
\arguments{
\item{fn.output}{The .h5 filename for the output, including folder directory}

\item{df.output}{A data.frame object with model results at each fixel, returned from FixelArray.lm() etc}

\item{analysis_name}{The subfolder name in results, holding the analysis results}

\item{overwrite}{If same analysis_name exists, whether overwrite (TRUE) or not (FALSE)}
}
\description{
Write outputs from fixel-based analysis out to the h5 file. Write one results (i.e. for one analysis) at a time. This is ".enh": 1) change to hdf5r; 2) write results with only one row for one fixel
debug tip: For "Error in H5File.open(filename, mode, file_create_pl, file_access_pl)", check if there is message 'No such file or directory'. Try absolute .h5 filename.
}
