% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelArray_Constructor.R
\name{writeResults}
\alias{writeResults}
\title{Write outputs from element-wise statistical analysis to the HDF5 file.}
\usage{
writeResults(
  fn.output,
  df.output,
  analysis_name = "myAnalysis",
  overwrite = TRUE
)
}
\arguments{
\item{fn.output}{A character, The HDF5 (.h5) filename for the output}

\item{df.output}{A data.frame object with element-wise statistical results, returned from `ModelArray.lm()` etc}

\item{analysis_name}{A character, the name of the results}

\item{overwrite}{If a group with the same analysis_name exists in HDF5 file, whether overwrite it (TRUE) or not (FALSE)}
}
\description{
Create a group named `analysis_name` in HDF5 file, then write the statistical results data.frame (i.e. for one analysis) in it.
}
\details{
debug tip: For "Error in H5File.open(filename, mode, file_create_pl, file_access_pl)", check if there is message 'No such file or directory'. Try absolute .h5 filename.
}
